+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/tool/editor/action/EncoderDecoderActionGroup.kt
@@ -9,14 +9,21 @@ import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.util.TextRange
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.EditorUtils.getSelectedText
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.EncoderDecoder
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.TransformationMode.DECODE
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.TransformationMode.ENCODE
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.encoderDecoders
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.executeTransformationInEditor
+import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.executeDecodingInEditor
+import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.editor.EncodersDecoders.executeEncodingInEditor
 
 internal open class EncoderDecoderActionGroup : DefaultActionGroup("Encoders/Decoders", false) {
   // -- Properties -------------------------------------------------------------------------------------------------- //
+
+  private val encoderActionGroup = createActionGroup(
+    title = "Encode To",
+    actions = EncodersDecoders.encoders.map { encoder -> EncoderAction(encoder) { getSourceText(it) } }
+  )
+  private val decoderActionGroup = createActionGroup(
+    title = "Decode To",
+    actions = EncodersDecoders.decoders.map { decoder -> DecoderAction(decoder) { getSourceText(it) } }
+  )
+
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exported Methods -------------------------------------------------------------------------------------------- //
 
@@ -25,10 +32,12 @@ internal open class EncoderDecoderActionGroup : DefaultActionGroup("Encoders/Dec
     e.presentation.isVisible = editor != null && editor.document.isWritable && getSourceText(e) != null
   }
 
-  final override fun getChildren(e: AnActionEvent?): Array<AnAction> = arrayOf(
-    EncoderDecoderActionsGroup(ENCODE) { getSourceText(it) },
-    EncoderDecoderActionsGroup(DECODE) { getSourceText(it) }
+  private val encoderDecoderActions: Array<AnAction> = arrayOf(
+    encoderActionGroup,
+    decoderActionGroup
   )
+
+  final override fun getChildren(e: AnActionEvent?): Array<AnAction> = encoderDecoderActions
 
   final override fun getActionUpdateThread() = ActionUpdateThread.BGT
 
@@ -38,32 +47,40 @@ internal open class EncoderDecoderActionGroup : DefaultActionGroup("Encoders/Dec
   }
 
   // -- Private Methods --------------------------------------------------------------------------------------------- //
+
+  private fun createActionGroup(title: String, actions: List<AnAction>) = object : DefaultActionGroup(title, true) {
+
+    private val decoderActions: Array<AnAction> = actions.toTypedArray()
+
+    override fun getChildren(e: AnActionEvent?): Array<AnAction> = decoderActions
+  }
+
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
-  private class EncoderDecodeAction(
-    val transformationMode: EncodersDecoders.TransformationMode,
-    val encoderDecoder: EncoderDecoder,
+  private class EncoderAction(
+    val encoder: EncodersDecoders.Encoder,
     val getSourceText: (AnActionEvent) -> Pair<String, TextRange>?
-  ) : DumbAwareAction(encoderDecoder.title, null, null) {
+  ) : DumbAwareAction(encoder.title, encoder.actionName, null) {
 
     override fun actionPerformed(e: AnActionEvent) {
       val editor = e.getData(EDITOR) ?: return
       val (text, textRange) = getSourceText(e) ?: return
-      executeTransformationInEditor(text, textRange, transformationMode, encoderDecoder, editor)
+      executeEncodingInEditor(text, textRange, encoder, editor)
     }
   }
 
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
-  private class EncoderDecoderActionsGroup(
-    val transformationMode: EncodersDecoders.TransformationMode,
+  private class DecoderAction(
+    val decoder: EncodersDecoders.Decoder,
     val getSourceText: (AnActionEvent) -> Pair<String, TextRange>?
-  ) : DefaultActionGroup(transformationMode.actionName, true) {
+  ) : DumbAwareAction(decoder.title, decoder.actionName, null) {
 
-    private val encoderDecoderActions: Array<AnAction> =
-      encoderDecoders.map { EncoderDecodeAction(transformationMode, it, getSourceText) }.toTypedArray()
-
-    override fun getChildren(e: AnActionEvent?): Array<AnAction> = encoderDecoderActions
+    override fun actionPerformed(e: AnActionEvent) {
+      val editor = e.getData(EDITOR) ?: return
+      val (text, textRange) = getSourceText(e) ?: return
+      executeDecodingInEditor(text, textRange, decoder, editor)
+    }
   }
 
   // -- Companion Object -------------------------------------------------------------------------------------------- //