+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - Add automatic input text case detection to the text case converter 
 - Add new tool: Text Statistic
-- Add support for dot text case
+- Add support for the "Dot Text Case"
+- Encoding in the editor now also supports commons hashing algorithms
 
 ### Changed